add delete and fix views file
