+++ management/static/management/delete-form.css
@@ -1,0 +1,62 @@
+@import url("https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;400;500;600;700;800&display=swap");
+
+body {
+  padding: 0;
+  margin: 0;
+  font-family: "Open Sans", sans-serif;
+  color: #37474f;
+  background-color: #f5f5f5;
+}
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+form {
+  margin: 5rem 12rem;
+  padding: 2rem;
+  border-radius: 1.5rem;
+  background-color: #fff;
+  margin-bottom: 5rem;
+}
+
+form .error {
+  background-color: #ffe0b2;
+  padding: 0.5rem 1rem;
+  border-radius: 0.5rem;
+  margin-bottom: 1rem;
+}
+
+form .inputs {
+  display: flex;
+  gap: 1rem;
+  margin-bottom: 1rem;
+}
+
+form input[type="submit"] {
+  display: block;
+  width: 100%;
+  padding: 1rem;
+  color: #f5f5f5;
+  font-weight: 600;
+  font-size: 1.2rem;
+  background-color: #d50000;
+  border: none;
+  border-radius: 0.5rem;
+  margin-bottom: 1rem;
+}
+
+form a {
+  display: block;
+  width: 100%;
+  padding: 1rem;
+  text-align: center;
+  color: #f5f5f5;
+  background-color: #43a047;
+  font-weight: 600;
+  font-size: 1.2rem;
+  text-decoration: none;
+  border-radius: 0.5rem;
+}
