+++ management/static/management/managed-planning.css
@@ -29,7 +29,8 @@ body {
   padding: 0px 5rem;
 }
 
-.week {
+.year,
+.padding {
   padding: 2rem;
   border-radius: 1.5rem;
   border: 2px solid #0002;
@@ -37,31 +38,33 @@ body {
   margin-bottom: 5rem;
 }
 
-.week h2 {
+.year .info,
+.padding .info {
   text-align: center;
-  margin-bottom: 2rem;
+  margin-bottom: 0.5rem;
+  font-weight: 500;
+  margin-bottom: 1rem;
 }
 
-.week .list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
+.year table,
+.padding table {
+  width: 100%;
+  border: 1px solid #0002;
+  border-collapse: collapse;
+  margin-bottom: 1rem;
 }
 
-.week .list .item {
-  min-width: 15rem;
-  padding: 1.5rem;
-  border-radius: 0.8rem;
-  background-color: #f5f5f5;
-  margin-bottom: 2rem;
-}
-
-.week .list .item p {
+.year table th,
+.year table td,
+.padding table th,
+.padding table td {
+  padding: 0.5rem;
   text-align: center;
-  font-weight: 600;
+  border: 1px solid #0002;
 }
 
-.week a {
+.year table td a,
+.padding table td a {
   display: block;
   text-decoration: none;
   text-align: center;
