+++ management/static/management/managed-promotion.css
@@ -29,7 +29,8 @@ body {
   padding: 0px 5rem;
 }
 
-.semester {
+.promotion,
+.subject {
   padding: 2rem;
   border-radius: 1.5rem;
   border: 2px solid #0002;
@@ -37,18 +38,21 @@ body {
   margin-bottom: 5rem;
 }
 
-.semester h2 {
+.promotion h2,
+.subject h2 {
   text-align: center;
   margin-bottom: 2rem;
 }
 
-.semester .list {
+.promotion .list,
+.subject .list {
   display: flex;
   flex-wrap: wrap;
   gap: 1rem;
 }
 
-.semester .list .item {
+.promotion .list .item,
+.subject .list .item {
   min-width: 15rem;
   padding: 1.5rem;
   border-radius: 0.8rem;
@@ -56,13 +60,21 @@ body {
   margin-bottom: 2rem;
 }
 
-.semester .list .item h3 {
+.promotion .list .item h3,
+.subject .list .item h3 {
   text-align: center;
   margin-bottom: 1rem;
   font-weight: 600;
 }
 
-.semester .list .item a {
+.promotion .list .item .btn,
+.subject .list .item .btn {
+  display: flex;
+  gap: 1rem;
+}
+
+.promotion .list .item .btn a,
+.subject .list .item .btn a {
   display: block;
   text-align: center;
   text-decoration: none;
@@ -74,7 +86,7 @@ body {
   background-color: #1e88e5;
 }
 
-.semester .list .item .weeks {
+.promotion .list .item .tps {
   display: flex;
   flex-wrap: wrap;
   justify-content: space-evenly;
@@ -82,13 +94,14 @@ body {
   margin-bottom: 1rem;
 }
 
-.semester .list .item .weeks .week {
+.promotion .list .item .tps .tp {
   border-radius: 0.5rem;
   padding: 0.5rem 1rem;
   background-color: #bbdefb;
 }
 
-.semester .btn-add {
+.promotion .btn-add,
+.subject .btn-add {
   display: block;
   text-decoration: none;
   text-align: center;
@@ -98,9 +111,4 @@ body {
   border-radius: 0.5rem;
   color: #f5f5f5;
   background-color: #1e88e5;
-  margin-bottom: 1rem;
-}
-
-.semester .btn-add:last-child {
-  margin-bottom: 0px;
 }
