+++ management/static/management/managed-semester.css
@@ -29,7 +29,7 @@ body {
   padding: 0px 5rem;
 }
 
-.td {
+.week {
   padding: 2rem;
   border-radius: 1.5rem;
   border: 2px solid #0002;
@@ -37,18 +37,18 @@ body {
   margin-bottom: 5rem;
 }
 
-.td h2 {
+.week h2 {
   text-align: center;
   margin-bottom: 2rem;
 }
 
-.td .list {
+.week .list {
   display: flex;
   flex-wrap: wrap;
   gap: 1rem;
 }
 
-.td .list .item {
+.week .list .item {
   min-width: 15rem;
   padding: 1.5rem;
   border-radius: 0.8rem;
@@ -56,12 +56,30 @@ body {
   margin-bottom: 2rem;
 }
 
-.td .list .item p {
+.week .list .item p {
   text-align: center;
   font-weight: 600;
+  margin-bottom: 1rem;
 }
 
-.td a {
+.week .list .item .btn {
+  display: flex;
+  gap: 1rem;
+}
+
+.week .list .item .btn a {
+  display: block;
+  text-align: center;
+  text-decoration: none;
+  color: #f5f5f5;
+  padding: 0.5rem;
+  border-radius: 0.5rem;
+  font-weight: 500;
+  width: 100%;
+  background-color: #1e88e5;
+}
+
+.week a {
   display: block;
   text-decoration: none;
   text-align: center;
