+++ management/static/management/managed-td.css
@@ -29,8 +29,7 @@ body {
   padding: 0px 5rem;
 }
 
-.year,
-.padding {
+.td {
   padding: 2rem;
   border-radius: 1.5rem;
   border: 2px solid #0002;
@@ -38,31 +37,51 @@ body {
   margin-bottom: 5rem;
 }
 
-.year .info,
-.padding .info {
+.td h2 {
   text-align: center;
-  margin-bottom: 0.5rem;
-  font-weight: 500;
+  margin-bottom: 2rem;
+}
+
+.td .list {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
+
+.td .list .item {
+  min-width: 15rem;
+  padding: 1.5rem;
+  border-radius: 0.8rem;
+  background-color: #f5f5f5;
+  margin-bottom: 2rem;
+}
+
+.td .list .item p {
+  text-align: center;
+  font-weight: 600;
   margin-bottom: 1rem;
 }
 
-.year table,
-.padding table {
-  width: 100%;
-  border: 1px solid #0002;
-  border-collapse: collapse;
-  margin-bottom: 1rem;
-}
-
-.year table th,
-.year table td,
-.padding table th,
-.padding table td {
-  padding: 0.5rem;
+.teacher .list .item a {
+  display: block;
   text-align: center;
-  border: 1px solid #0002;
-}
-
-.padding table td a {
   text-decoration: none;
+  color: #f5f5f5;
+  padding: 0.5rem;
+  border-radius: 0.5rem;
+  font-weight: 500;
+  width: 100%;
+  background-color: #1e88e5;
+}
+
+.td a {
+  display: block;
+  text-decoration: none;
+  text-align: center;
+  font-size: 1.2rem;
+  font-weight: 500;
+  padding: 0.5rem;
+  border-radius: 0.5rem;
+  color: #f5f5f5;
+  background-color: #1e88e5;
 }
