+++ management/templates/management/add-week.html
@@ -7,7 +7,7 @@
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/datepicker/0.6.5/datepicker.min.css" integrity="sha256-b88RdwbRJEzRx95nCuuva+hO5ExvXXnpX+78h8DjyOE=" crossorigin="anonymous" />
 <script src="https://cdnjs.cloudflare.com/ajax/libs/datepicker/0.6.5/datepicker.min.js" integrity="sha256-/7FLTdzP6CfC1VBAj/rsp3Rinuuu9leMRGd354hvk0k=" crossorigin="anonymous"></script>
 
-<link rel="stylesheet" type="text/css" href="{% static 'management/form.css' %}" />
+<link rel="stylesheet" type="text/css" href="{% static 'management/add-form.css' %}" />
 
 <form action="{% url 'management:add-week' year_id semester_id%}" method="post">
   {% csrf_token %}
@@ -28,6 +28,5 @@
   <!---->
   <div class="inputs">{{ form }}</div>
   <input type="submit" value="Submit" />
-  <a href="{% url 'management:edit-semester' year_id semester_id%}"><p>Annuler</p></a>
+  <a href="{% url 'management:managed-semester' year_id semester_id%}"><p>Annuler</p></a>
 </form>
-g
