+++ management/templates/management/delete-form.html
@@ -1,6 +1,6 @@
 {% load static %}
 
-<link rel="stylesheet" type="text/css" href="{% static 'management/add-form.css' %}" />
+<link rel="stylesheet" type="text/css" href="{% static 'management/delete-form.css' %}" />
 
 <form action="{{ post_url }}" method="post">
   {% csrf_token %}
@@ -12,6 +12,6 @@
   {% endif %}
   <!---->
   <div class="inputs">{{ form }}</div>
-  <input type="submit" value="Submit" />
-  <a href="{{ back_url }}"><p>Annuler</p></a>
+  <input type="submit" value="Delete" />
+  <a href="{{ back_url }}">Annuler</a>
 </form>
