+++ management/templates/management/managed-promotion.html
@@ -1,0 +1,49 @@
+{% load static %}
+
+<link rel="stylesheet" type="text/css" href="{% static 'management/managed-promotion.css' %}" />
+
+<div class="title">
+  <h1>{{ promotion.name_promotion }}</h1>
+  <p>{{ promotion.year.name_year }}</p>
+  <a class="btn-back" href="{% url 'management:index' %}">← back</a>
+</div>
+<div class="container">
+  <div class="promotion">
+    <h2>Promotion</h2>
+    <div class="list">
+      {% for one_td in td %}
+      <div class="item">
+        <h3>{{ one_td }}</h3>
+        <dev class="tps">
+          {% for one_tp in one_td.tp_set.all %}
+          <div class="tp">
+            <p>{{ one_tp }}</p>
+          </div>
+          {% endfor %}
+        </dev>
+        <div class="btn">
+          <a href="{% url 'management:managed-td' promotion.id one_td.id %}">Edit</a>
+          <a href="{% url 'management:delete-td' promotion.id one_td.id %}">Delete</a>
+        </div>
+      </div>
+      {% endfor %}
+    </div>
+    <a class="btn-add" href="{% url 'management:add-td' promotion.id %}">Ajouter un TD</a>
+  </div>
+
+  <div class="subject">
+    <h2>Matiere</h2>
+    <div class="list">
+      {% for one_subject in subject %}
+      <div class="item">
+        <h3>{{ one_subject }}</h3>
+        <div class="btn">
+          <a href="{% url 'management:managed-subject' promotion.id one_subject.id %}">Edit</a>
+          <a href="{% url 'management:delete-subject' promotion.id one_subject.id %}">Delete</a>
+        </div>
+      </div>
+      {% endfor %}
+    </div>
+    <a class="btn-add" href="{% url 'management:add-subject' promotion.id %}">Ajouter une matiere</a>
+  </div>
+</div>
