+++ management/templates/management/managed-semester.html
@@ -1,11 +1,11 @@
 {% load static %}
 
-<link rel="stylesheet" type="text/css" href="{% static 'management/edit-semester.css' %}" />
+<link rel="stylesheet" type="text/css" href="{% static 'management/managed-semester.css' %}" />
 
 <div class="title">
   <h1>{{ semester.name_semester }}</h1>
   <p>{{ year.name_year }}</p>
-  <a class="btn-back" href="{% url 'management:edit-year' year.id%}">← back</a>
+  <a class="btn-back" href="{% url 'management:managed-year' year.id%}">← back</a>
 </div>
 
 <div class="container">
@@ -15,6 +15,7 @@
       {% for one_week in semester.week_set.all %}
       <div class="item">
         <p>{{ one_week }}</p>
+        <a href="{% url 'management:delete-week' year.id semester.id one_week.id %}">Delete</a>
       </div>
       {% endfor %}
     </div>
