+++ management/templates/management/managed-subject.html
@@ -1,11 +1,11 @@
 {% load static %}
 
-<link rel="stylesheet" type="text/css" href="{% static 'management/edit-subject.css' %}" />
+<link rel="stylesheet" type="text/css" href="{% static 'management/managed-subject.css' %}" />
 
 <div class="title">
   <h1>{{ subject.name_subject }}</h1>
   <p>{{ promotion.name_promotion }}</p>
-  <a href="{% url 'management:edit-promotion' promotion.id%}">← back</a>
+  <a href="{% url 'management:managed-promotion' promotion.id %}">← back</a>
 </div>
 <div class="container">
   <div class="cm">
@@ -18,6 +18,7 @@
         <th>Nom professeur</th>
         <th>Nombre de séances</th>
         <th>Semaine</th>
+        <th>Action</th>
       </tr>
       {% for one_sessions in subject.sessions_set.all %}
       <tr>
@@ -25,12 +26,13 @@
         <td>{{ one_sessions.teacher }}</td>
         <td>{{ one_sessions.number_sessions }}</td>
         <td>{{ one_sessions.week }}</td>
+        <td><a href="{% url 'management:delete-session' promotion.id subject.id one_sessions.id %}">Delete</a></td>
         {% endif %}
       </tr>
       {% endfor %}
     </table>
     {% if nb_sessions.cm.remaining > 0%}
-    <a href="{% url 'management:add-cm-session' promotion.id subject.id%}">Ajouter un CM</a>
+    <a href="{% url 'management:add-cm-session' promotion.id subject.id %}">Ajouter un CM</a>
     {% endif %}
   </div>
 
@@ -48,19 +50,21 @@
         <th>Nom professeur</th>
         <th>Nombre de séances</th>
         <th>Semaine</th>
+        <th>Action</th>
       </tr>
       {% for one_sessions in list_session %}
       <tr>
         <td>{{ one_sessions.teacher }}</td>
         <td>{{ one_sessions.number_sessions }}</td>
         <td>{{ one_sessions.week }}</td>
+        <td><a href="{% url 'management:delete-session' promotion.id subject.id one_sessions.id %}">Delete</a></td>
       </tr>
       {% endfor %}
     </table>
     {% endfor %}
     <!---->
     {% if nb_sessions.td.remaining > 0%}
-    <a href="{% url 'management:add-td-session' promotion.id subject.id%}"><p>Ajouter un TD</p></a>
+    <a href="{% url 'management:add-td-session' promotion.id subject.id %}"><p>Ajouter un TD</p></a>
     {% endif %}
   </div>
 
@@ -78,19 +82,21 @@
         <th>Nom professeur</th>
         <th>Nombre de séances</th>
         <th>Semaine</th>
+        <th>Action</th>
       </tr>
       {% for one_sessions in list_session %}
       <tr>
         <td>{{ one_sessions.teacher }}</td>
         <td>{{ one_sessions.number_sessions }}</td>
         <td>{{ one_sessions.week }}</td>
+        <td><a href="{% url 'management:delete-session' promotion.id subject.id one_sessions.id %}">Delete</a></td>
       </tr>
       {% endfor %}
     </table>
     {% endfor %}
     <!---->
     {% if nb_sessions.tp.remaining > 0%}
-    <a href="{% url 'management:add-tp-session' promotion.id subject.id%}"><p>Ajouter un TP</p></a>
+    <a href="{% url 'management:add-tp-session' promotion.id subject.id %}"><p>Ajouter un TP</p></a>
     {% endif %}
   </div>
 </div>
