+++ management/templates/management/managed-teacher.html
@@ -1,6 +1,6 @@
 {% load static %}
 
-<link rel="stylesheet" type="text/css" href="{% static 'management/edit-teacher.css' %}" />
+<link rel="stylesheet" type="text/css" href="{% static 'management/managed-teacher.css' %}" />
 
 <div class="title">
   <h1>{{ teacher.last_name }}</h1>
